+++ backend/mora/settings.py
@@ -133,6 +133,7 @@ app_config = {
     "SAML_AUTH_ENABLE": config["saml_sso"]["enable"],
     "SAML_IDP_INSECURE": config["saml_sso"]["idp_insecure"],
     "SAML_FORCE_HTTPS": config["saml_sso"]["force_https"],
+    "SAML_SP_DOMAIN": config["saml_sso"]["sp_domain"],
     "SAML_NAME_ID_FORMAT": config["saml_sso"]["name_id_format"],
     "SAML_WANT_NAME_ID": config["saml_sso"]["want_name_id"],
     "SAML_WANT_ATTRIBUTE_STATEMENT":