+++ NEWS.rst
@@ -7,6 +7,7 @@ Bug fixes
 ---------
 
 * #35531: Fix org unit rename dialog error handling
+* #35897: Fix conf_db health endpoint not catching certain errors
 
 
 Version 1.6.0, 2020-03-24